[ PHASE ONE ] Add inStockOnly handler and rename folder with mocked data
