+++ src/thinking-in-react/App.js
@@ -1,5 +1,5 @@
 import "./App.css";
-import { PRODUCTS } from "./data/products";
+import { PRODUCTS } from "./mock/products";
 import FilterableProductTable from "./filterable-product-table/FilterableProductTable";
 
 export default function App() {
