+++ src/thinking-in-react/mock/products.js
